docs: embed docs to documentation
Use one example file and embed the examples to godoc
