+++ examples/main.go
@@ -1,0 +1,69 @@
+// Copyright (c) 0x6flab. All rights reserved.
+//
+// SPDX-License-Identifier: GNU GENERAL PUBLIC LICENSE
+
+package main
+
+import (
+	"fmt"
+
+	"github.com/0x6flab/namegenerator"
+)
+
+func main() {
+	generator := namegenerator.NewNameGenerator()
+	name := generator.Generate()
+	fmt.Println(name)
+
+	names := generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithGender(namegenerator.Male)
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithGender(namegenerator.Female)
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithGender(namegenerator.NonBinary)
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithPrefix("Dr. ")
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithSuffix(" Jr.")
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithPrefix("Dr. ").WithSuffix(" Jr.")
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+
+	generator = namegenerator.NewNameGenerator().WithGender(namegenerator.Male).WithPrefix("Dr. ").WithSuffix(" Jr.")
+	name = generator.Generate()
+	fmt.Println(name)
+
+	names = generator.GenerateNames(10)
+	fmt.Println(names)
+}
