--- examples/prefix/main.go
@@ -1,17 +1,0 @@
-// Copyright (c) 0x6flab. All rights reserved.
-//
-// SPDX-License-Identifier: GNU GENERAL PUBLIC LICENSE
-
-package main
-
-import (
-	"fmt"
-
-	"github.com/0x6flab/namegenerator"
-)
-
-func main() {
-	generator := namegenerator.NewNameGenerator().WithPrefix("Dr. ")
-	name := generator.Generate()
-	fmt.Println(name)
-}
